requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.4.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==39.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
et-xmlfile==1.1.0
fastapi==0.94.1
filelock==3.9.1
Flask==2.2.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.112
layoutparser==0.0.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
openai==0.27.2
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
Pillow==9.4.0
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
pymongo==4.3.3
pypandoc==1.11
pypdf==3.5.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
slack-bolt==1.16.4
slack-sdk==3.20.2
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.4
unstructured-inference==0.2.4
urllib3==1.26.15
uvicorn==0.21.0
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
